+++ app/routes/components/Typography/changelog.codegen.ts
@@ -22,4 +22,5 @@ export const changelog = ''
   + '- 2025-11-17: Remove `Typography.Root` component. Just use `Typography` as the\n'
   + '  root component. Some minor styles are updated.\n'
   + '- 2025-12-07: Update the padding for `Typography.P`.\n'
+  + '- 2025-12-12: Add `Typography.A` component.\n'
   ;